pyproject.toml
[tool.poetry]
name = "repl-nix-llama-query-vm"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-cors = "^5.0.0"
requests = "^2.32.3"
google-cloud-language = "^2.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
